Tilføjet login på UI

+++ backend/kodeordsmanager/src/kodeordsmanager.domain/Models/JwtModel.cs
@@ -6,5 +6,5 @@ public class JwtModel
     public required string Key { get; init; }
     public required string Issuer { get; init; }
     public required string Audience { get; init; }
-    public required double DurationInMinutes { get; init; }
+    public required int DurationInMinutes { get; init; }
 }
